+++ example/lib/ui_widget_gallery/imageview_page.dart
@@ -14,20 +14,20 @@ class ImageViewPage extends StatelessWidget {
       body: VStack([
         '24px'.text.bold.make().pOnly(bottom: 12),
         HStack([
-          ExImageView(url: '', size: 24).pOnly(right: 12),
-          ExImageView(url: '', size: 24).pOnly(right: 12),
+          ExImageView(source: '', size: 24).pOnly(right: 12),
+          ExImageView(source: '', size: 24).pOnly(right: 12),
           ExImageView(
-                  url:
+                  source:
                       'https://picsum.photos/id/${random.integer(999, min: 100)}/200/300',
                   size: 24)
               .pOnly(right: 12),
           ExImageView(
-                  url:
+                  source:
                       'https://picsum.photos/id/${random.integer(999, min: 100)}/200/300',
                   size: 24)
               .pOnly(right: 12),
           ExImageView(
-                  url:
+                  source:
                       'https://picsum.photos/id/${random.integer(999, min: 100)}/200/300',
                   size: 24)
               .pOnly(right: 12),
@@ -35,20 +35,20 @@ class ImageViewPage extends StatelessWidget {
         Divider().pSymmetric(v: 16),
         '32px'.text.bold.make().pOnly(bottom: 12),
         HStack([
-          ExImageView(url: '', size: 32).pOnly(right: 12),
-          ExImageView(url: '', size: 32).pOnly(right: 12),
+          ExImageView(source: '', size: 32).pOnly(right: 12),
+          ExImageView(source: '', size: 32).pOnly(right: 12),
           ExImageView(
-                  url:
+                  source:
                       'https://picsum.photos/id/${random.integer(999, min: 100)}/200/300',
                   size: 32)
               .pOnly(right: 12),
           ExImageView(
-                  url:
+                  source:
                       'https://picsum.photos/id/${random.integer(999, min: 100)}/200/300',
                   size: 32)
               .pOnly(right: 12),
           ExImageView(
-                  url:
+                  source:
                       'https://picsum.photos/id/${random.integer(999, min: 100)}/200/300',
                   size: 32)
               .pOnly(right: 12),
@@ -57,23 +57,23 @@ class ImageViewPage extends StatelessWidget {
         '40px'.text.bold.make().pOnly(bottom: 12),
         HStack([
           ExImageView(
-                  url: 'assets/images/ic_dialog_error.svg',
+                  source: 'assets/images/ic_dialog_error.svg',
                   size: 40,
                   package: 'ex')
               .pOnly(right: 12),
-          ExImageView(url: '', size: 40).pOnly(right: 12),
+          ExImageView(source: '', size: 40).pOnly(right: 12),
           ExImageView(
-                  url:
+                  source:
                       'https://picsum.photos/id/${random.integer(999, min: 100)}/200/300',
                   size: 40)
               .pOnly(right: 12),
           ExImageView(
-                  url:
+                  source:
                       'https://picsum.photos/id/${random.integer(999, min: 100)}/200/300',
                   size: 40)
               .pOnly(right: 12),
           ExImageView(
-                  url:
+                  source:
                       'https://picsum.photos/id/${random.integer(999, min: 100)}/200/300',
                   size: 40)
               .pOnly(right: 12),
@@ -81,20 +81,20 @@ class ImageViewPage extends StatelessWidget {
         Divider().pSymmetric(v: 16),
         '48px'.text.bold.make().pOnly(bottom: 12),
         HStack([
-          ExImageView(url: '', size: 48).pOnly(right: 12),
-          ExImageView(url: '', size: 48).pOnly(right: 12),
+          ExImageView(source: '', size: 48).pOnly(right: 12),
+          ExImageView(source: '', size: 48).pOnly(right: 12),
           ExImageView(
-                  url:
+                  source:
                       'https://picsum.photos/id/${random.integer(999, min: 100)}/200/300',
                   size: 48)
               .pOnly(right: 12),
           ExImageView(
-                  url:
+                  source:
                       'https://picsum.photos/id/${random.integer(999, min: 100)}/200/300',
                   size: 48)
               .pOnly(right: 12),
           ExImageView(
-                  url:
+                  source:
                       'https://picsum.photos/id/${random.integer(999, min: 100)}/200/300',
                   size: 48)
               .pOnly(right: 12),
@@ -102,20 +102,20 @@ class ImageViewPage extends StatelessWidget {
         Divider().pSymmetric(v: 16),
         '56px'.text.bold.make().pOnly(bottom: 12),
         HStack([
-          ExImageView(url: '', size: 56).pOnly(right: 12),
-          ExImageView(url: '', size: 56).pOnly(right: 12),
+          ExImageView(source: '', size: 56).pOnly(right: 12),
+          ExImageView(source: '', size: 56).pOnly(right: 12),
           ExImageView(
-                  url:
+                  source:
                       'https://picsum.photos/id/${random.integer(999, min: 100)}/200/300',
                   size: 56)
               .pOnly(right: 12),
           ExImageView(
-                  url:
+                  source:
                       'https://picsum.photos/id/${random.integer(999, min: 100)}/200/300',
                   size: 56)
               .pOnly(right: 12),
           ExImageView(
-                  url:
+                  source:
                       'https://picsum.photos/id/${random.integer(999, min: 100)}/200/300',
                   size: 56)
               .pOnly(right: 12),